code/src/aidhp-naruto-backend/graphCreator/views.py
import os
import google.generativeai as genai
import numpy as np
import pandas as pd
from django.http import JsonResponse
from django.views.decorators.csrf import csrf_exempt
from django.views.decorators.http import require_http_methods

# Configure Gemini API
genai.configure(api_key=os.getenv('GEMINI_API_KEY'))

def generate_market_trend_data():
    """
    Generate market trend data using Gemini 2.0 Flash
    """
    model = genai.GenerativeModel('gemini-2.0-flash')
    
    prompt = """
    Generate a detailed financial market trend analysis for the past 12 months.
    Provide data points for:
    1. Projected market growth rate
    2. Sector performance variations
    3. Potential investment opportunities
    4. Risk assessment metrics

    Respond with a structured JSON format containing numerical data.
    """

    try:
        response = model.generate_content(prompt)
        
        # Fallback to default data if generation fails
        if not response.text:
            return {
                'time_series': [
                    {'month': 'Jan', 'value': 45.2},
                    {'month': 'Feb', 'value': 47.6},
                    {'month': 'Mar', 'value': 46.8},
                    {'month': 'Apr', 'value': 48.3},
                    {'month': 'May', 'value': 49.1},
                    {'month': 'Jun', 'value': 50.5},
                    {'month': 'Jul', 'value': 51.2},
                    {'month': 'Aug', 'value': 52.7},
                    {'month': 'Sep', 'value': 53.4},
                    {'month': 'Oct', 'value': 54.9},
                    {'month': 'Nov', 'value': 56.2},
                    {'month': 'Dec', 'value': 57.6}
                ],
                'growth_rate': 27.5,
                'risk_level': 'moderate'
            }

        # Parse and validate the response
        try:
            parsed_data = json.loads(response.text)
            return parsed_data
        except json.JSONDecodeError:
            # Generate default data if parsing fails
            return {
                'time_series': [
                    {'month': 'Jan', 'value': 45.2},
                    # ... other months ...
                ],
                'growth_rate': 27.5,
                'risk_level': 'moderate'
            }

    except Exception as e:
        print(f"Error generating market trends: {e}")
        return {
            'time_series': [
                {'month': 'Jan', 'value': 45.2},
                # ... other months ...
            ],
            'growth_rate': 27.5,
            'risk_level': 'moderate'
        }

@csrf_exempt
@require_http_methods(["POST"])
def graph_gen_view(request):
    """
    API endpoint for generating market trend graph data
    """
    try:
        market_trends = generate_market_trend_data()
        return JsonResponse({
            'status': 'success',
            'market_trends': market_trends
        }, status=200)
    except Exception as e:
        return JsonResponse({
            'status': 'error',
            'message': str(e)
        }, status=500)
